double to int

+++ Core_CPP/csharp_connector.h
@@ -2,6 +2,6 @@
 
 
 extern "C" {
-	__declspec(dllexport) int* createRoom(double* coords, int* roomList_ptr[], int n, bool snap_mode);
-	__declspec(dllexport) int* updateRoom(int* room, double* coords, int n, bool snap_mode);
+	__declspec(dllexport) int* createRoom(int* coords, int* roomList_ptr[], int n, bool snap_mode);
+	__declspec(dllexport) int* updateRoom(int* room, int* coords, int n, bool snap_mode);
 }
